==> analysis/00_zora_download.Rmd <==
---
title: "01_zora_download"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Generating ZORA JSON files from API

These have been run in advance and then gzipped to save space ..

```{r eval=TRUE}
stub <- "wget -q -O data/zora_YEAR.js --no-check-certificate https://www.zora.uzh.ch/cgi/exportview/yearnew/YEAR/JSON/YEAR.js"

z <- 2021:2001

for(i in z) {
  cmd <- gsub("YEAR",i,stub)
  cat(cmd,"\n")
  system(cmd)
}
system("gzip data/*.js")
```

## sessionInfo()

```{r sessioninfo}
sessionInfo()
```
