==> analysis/02_zora_plots.Rmd <==
---
title: "02_zora_plots"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load packages

```{r cars}
library(dplyr)
library(ggplot2)
```


## Subjects

```{r subjects}

my_scale <- scale_fill_manual(values=c("gray48","darkorange1","chartreuse4","gold"))


outdir <- "output"
tbl_subjects <- readRDS(file.path(outdir, "tbl_subjects.rds"))
tbl_eprints <- readRDS(file.path(outdir, "tbl_eprints.rds"))


df <- tbl_subjects %>% left_join(tbl_eprints)
df$oa_status <- factor(df$oa_status, levels=c("closed","hybrid","green","gold"))

nm <- names(sort(table(df$name), decreasing = TRUE)[1:50])

ggplot(df %>% filter(date > 2007, date < 2020, name %in% "Institute of Molecular Life Sciences"), aes(x=date, fill=oa_status)) + 
  geom_bar() + 
  facet_wrap(~name, scales="free_y", ncol=5) +
  theme(axis.text.x = element_text(angle = 90)) +
  my_scale


df <- df %>% filter(date > 2007, date < 2021, 
                    !(df$parent_name %in% c("National licences", 
                                       "University of Zurich")))

# df1 <- df[grepl("^0[1-8] [FRV]", df$parent_name),]


df1 <- df[df$parent_name=="07 Faculty of Science",]


# nm <- names(sort(table(df$parent_name), decreasing = TRUE)[1:15])

ggplot(df1 %>% filter(!is.na(date), !(name=="Institute of Zoology (former)")), aes(x=date, fill=oa_status)) + 
  geom_bar() + 
  facet_wrap(~name, scales="free_y", ncol=3) +
  theme(axis.text.x = element_text(angle = 90)) +
  xlab("year") + ylab("count") +
  my_scale

ggsave("ZORA_OA_numbers_MNF_by_department.pdf", width=12, height=9)


df1 <- df1[df1$parent_name=="07 Faculty of Science",]
# df1 <- na.omit(df1)

ggplot(df1, aes(x=date, fill=oa_status)) + 
  geom_bar() + 
  facet_wrap(~name, scales="free_y", ncol=5) +
  theme(axis.text.x = element_text(angle = 90)) +
  my_scale


df2 <- df1 %>% group_by(name) %>% 
  summarize(sum_oa=sum(oa_status!="closed"), n=n()) %>%
  mutate(mean_oa=sum_oa/n) %>% 
  filter(n>100) %>% 
  arrange(-mean_oa)

top_bottom <- rbind( top_n(df2, 4), top_n(df2, -4) )

df3 <- df[df$name %in% top_bottom$name,]
df3$name <- factor(df3$name, levels=top_bottom$name)

ggplot(df3, aes(x=date, fill=oa_status)) + 
  geom_bar() + 
  facet_wrap(~name, scales="free_y", ncol=4) +
  theme(axis.text.x = element_text(angle = 90)) +
  my_scale

ggsave("top_bottom4.pdf", width=12, height=6)

```

## Dewey

```{r dewey}
# read in Dewey file to get classifications
dewey_file <- read.table("https://www.gutenberg.org/cache/epub/12513/pg12513.txt", 
                         sep="\n", quote="")
# dewey_list <- grep("^[0-9][0-9]0", dewey_file$V1, value=TRUE)
dewey_list <- grep("^=[0-9][0-9][0-9]", dewey_file$V1, value=TRUE)
dewey_list <- gsub("=","", dewey_list)
dewey_num <- substr(dewey_list, 1, 3)
dewey_cat <- trimws(substr(dewey_list, 5, 200))
dewey_lookup <- setNames(dewey_cat, dewey_num)
rm(dewey_file, dewey_list, dewey_num, dewey_cat)

tbl_dewey <- readRDS(file.path(outdir, "tbl_dewey.rds"))


top_dewey <- names(sort(table(tbl_dewey$dewey), decreasing = TRUE)[1:18])
tbl_dewey <- tbl_dewey %>% dplyr::filter(dewey %in% top_dewey)
tbl_dewey$field <- dewey_lookup[gsub("ddc","", tbl_dewey$dewey)]


df <- tbl_dewey %>% left_join(tbl_eprints)

# ggplot(mydf, aes(x=date, fill=oa_status)) + 
#   geom_bar() + facet_wrap(~oa_status, ncol=1, scales = "free_y")
# ggplot(df, aes(x=date, fill=oa_status)) + 
#   geom_bar() + scale_y_sqrt()

df$oa_status <- factor(df$oa_status, levels=c("closed","hybrid","green","gold"))

ggplot(df %>% filter(date > 2007, date < 2020), aes(x=date, fill=oa_status)) + 
  geom_bar() + 
  facet_wrap(~paste0(gsub("ddc","",dewey),":",field), scales="free_y", ncol=6) +
  theme(axis.text.x = element_text(angle = 90)) +
  my_scale

ggsave("by_dewey.pdf", width=15, height=7)

```

## ORCID

```{r orcid}

tbl_orcid <- readRDS(file.path(outdir, "tbl_orcid.rds"))

df <- tbl_orcid %>% left_join(tbl_eprints) %>% 
  left_join(tbl_subjects)

df <- df[df$name == "Institute of Molecular Life Sciences",]

df$oa_status <- factor(df$oa_status, levels=c("closed","hybrid","green","gold"))

ggplot(df %>% filter(date>2007, date<2020), aes(x=date, fill=oa_status)) + 
  geom_bar() + 
  theme(axis.text.x = element_text(angle = 90)) +
  my_scale


  
s```
